Introduce the Stopwatch class
The Stopwatch class is in charge of keeping track of time and any
time-related events. The reason for having it be a separate class is
because it can be easily tested in isolation while providing meaningful
methods for the rest of the application. In exchange, it can be mocked
at will and injected into other classes to make it simpler to simulate
how time passes instead of having to play with Timecop or similar gems.