Allow logger and redis to be configured globally
It doesn't seem like it'll be very useful to allow users to configure
the logger and redis for every Unique Thread, so they can be globally
configured now. This should be specially convenient for Rails app, where
it can be configured once on an initializer and then used liberally
wherever.