Allow users to configure their own error reporting tools
When an exception is raised within a thread, it doesn't bubble up to the
parent thread or anything like that, so it's easy to not notice those
exceptions. Because of this, UniqueThread now allows users to add as
many error reporting tools they need and know that if their code breaks
they'll know.